Add table with project specific libraries
KiCAD libraries are ongoing development and subject to change. To be able
to review schematics against a fix version of the libraries the libraries
used in the schematic are added to the project as git submodules.

To configure KiCAD to use these libraries a configuration file is needed.
Add the configuration file "sym-lib-table".

Signed-off-by: Manuel Traut <manut@linutronix.de>
[enhance commit message]
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>