Change GPIO resistors value
The A20 GPIOs are capable of a current of ~20 mA. The 100 Ohms resistor
to the opto coupler limits the current to ~20 mA. Driving the GPIOs at
current limit may cause thermal failures. To be on save side change the
value to 200 Ohms since the opto works fine with 10 mA also.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>