Switch PE plane from bottom to top
The high voltage part on bottom PCB must be secured by some isolator
anyway. On the top side the user need to handle with screw drivers.
Put the high voltage routes on the bottom side and switch the PE plan
to the top side. While at it, remove the now unneeded cuttings.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>